Introduced Celeritas framework
A new package, Celeritas, has been added to the project. This includes a main struct for the framework and two key functions: New and Init. The 'New' function creates an initial directory structure for a new Celeritas project while 'Init' checks if folders exist in the root path and creates them if they don't.