Added directory creation helper function
A new helper function, CreateDirIfNotExist, has been added to the Celeritas package. This function checks if a directory exists at a given path and creates it with specified permissions if it doesn't exist. It returns nil if the directory already exists or is successfully created, otherwise it returns an error.