Introduced Celeritas initialization
A new function, initApplication, has been added to initialize the Celeritas library. This function retrieves the current working directory and uses it to create a new instance of Celeritas. If any errors occur during these processes, they are logged and cause the application to exit. The initialized Celeritas instance is then wrapped in an application struct and returned.