Updated dependencies and module path
The go.mod file has been updated to replace the celeritas dependency with a local version. The required version of celeritas has also been bumped up from v1.0.3 to v1.0.4. Additionally, the go-chi/chi/v5 package was added as an indirect dependency in the project's module requirements.

The go.sum file was updated accordingly to reflect these changes in dependencies.