Updated Celeritas package version
The Celeritas package has been updated from v1.0.7 to v1.0.9 in the go.mod file and other relevant places in the codebase. This update also includes changes to the createRender function, which no longer takes a parameter but uses the receiver's properties directly instead. The README.md file was also slightly modified for better readability of a link description.