Added new home and base layout views
Two new view templates have been added to the application. The 'home.jet' file is a new homepage template that includes a welcome message and an image. The 'base.jet' file is a basic layout template that includes meta tags, links to favicon images, and Bootstrap CSS. It also has placeholders for custom CSS, page content, and JavaScript which can be filled in by other templates extending it.